requirements.txt
langchain==0.3.25
langchain-groq==0.2.4
langgraph==0.3.34
chromadb==0.6.3
sentence-transformers==3.4.1
python-dotenv==1.1.0
